Extract fake watch settings UI and make state observable
- Move fakeWatchItems and FakeWatchPickerDialog to coredevices.pebble.fake package

- Add StateFlow-based fakeWatches/activeFakeWatch to FakeWatchConfigStore

- Collect fake watch state in WatchSettingsScreen and pass to fakeWatchItems

- Fix UI not updating when removing fake watches
